add rollback capabilities, only used if verifying
it's only used if we're given training data from a previous round, which
should currently be impossible as verifying isnt impl'd.

we also unify step as u32 everywhere.

+++ shared/client/src/protocol.rs
@@ -8,7 +8,7 @@ pub type NE = NetworkEvent<BroadcastMessage, Payload>;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct BroadcastMessage {
-    pub step: u64,
+    pub step: u32,
     pub batch_id: u64,
     pub commitment: Commitment,
     pub ticket: BlobTicket,
@@ -17,7 +17,7 @@ pub struct BroadcastMessage {
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Payload {
-    pub step: u64,
+    pub step: u32,
     pub batch_id: u64,
     pub distro_results: Vec<SerializedDistroResult>,
 }
